vm: retire the HIR walker — MIR is the only VM codegen path (#252 Phase 6)
Across #334–#338 the MIR walker took over every VM-compile shape: fn
bodies (entry + dependency modules), first-class fn values, built-in
records, discard bindings, and — most recently — top-level module
statements. The HIR walker (`compile_expr` + the call / pattern / record
emitters) survived only as a per-fn fallback and a pure-HIR entry point.
This deletes it.

Codegen:
  * the per-fn dispatch (entry + dep modules) and `compile_top_level` are
    now fail-soft: a fn / statement that doesn't lower to MIR is a hard
    `CompileError` (only reachable on malformed / typecheck-rejected
    input — the full corpus + suite lower with zero rejections), not a
    silent fall-back to a second walker.
  * `compile_program` (the pure-HIR entry) and the `use_mir` flag are
    gone; every entry point takes the single MIR path.
  * deleted `compile_fn` / `compile_fn_with_scope` / `compile_body` /
    `compile_stmt` / `compile_expr` and the whole `calls.rs` (HIR call +
    ctor + leaf-op emitter) and `patterns.rs` (HIR match emitter). The
    seven small helpers the MIR walker still reuses (canonical fn / type
    names, id-by-name resolution, literal `NanValue`s, buffer-intrinsic
    opcodes) moved to a focused `resolve_helpers.rs`; `compile_literal` /
    `compile_ident` stay in `expr.rs`.
  * dropped now-dead `SymbolTable::resolve_symbol_ref` /
    `resolve_namespace_path` and the `CallTarget` enum + arm-slot
    pattern-binding helpers.

Net ~2160 lines removed. There is now one VM codegen path, not two — the
duplication MIR was meant to kill is gone.

Tests:
  * the inline VM codegen tests + `vm_spec` execute through the MIR path;
    one `vm_spec` case that bound a builtin to a local (`f = String.len`,
    which the type checker rejects and only the HIR walker tolerated) was
    rewritten to a valid program. The eval_spec cross-module ctor case
    was resolved up front in #337.
  * full suite green; clippy --features wasm -D warnings clean.
